add if docker exists check

+++ rice-docker/rice-docker.sh
@@ -11,39 +11,43 @@ function install_docker {
             sudo echo " > sudo OK";
         fi
 
-    	echo " > Making sure curl is installed";
-    	if sudo apt install -y curl ; then
-    		echo " > curl installed";
-    	else
-    		echo " /!\ Something went wrong with curl installation.";
-    	fi
+        if [ -f /usr/bin/docker ] ; then
+            echo " > Making sure curl is installed";
+            if sudo apt install -y curl ; then
+                echo " > curl installed";
+            else
+                echo " /!\ Something went wrong with curl installation.";
+            fi
 
-        echo " > Grabbing docker install script from docker.com";
-        if curl -sSL https://get.docker.com/ -o install-docker.sh ; then
-        	echo " > install-docker.sh downloaded";
+            echo " > Grabbing docker install script from docker.com";
+            if curl -sSL https://get.docker.com/ -o install-docker.sh ; then
+                echo " > install-docker.sh downloaded";
+            else
+                echo " > failed to download install-docker.sh";
+            fi
+
+            echo " > Making install-docker.sh executable";
+            if chmod +x ./install-docker.sh ; then
+                echo " > install-docker.sh now executable";
+            else
+                echo " /!\ install-docker.sh failed to change permission";
+            fi
+
+            echo " > Executing install-docker.sh";
+            if ./install-docker.sh ; then
+                echo " > install-docker.sh complete";
+            else 
+                echo " /!\ install-docker.sh failed to run correctly";
+            fi
+
+            echo " > Cleaning up install-docker.sh";
+            if rm -f ./install-docker.sh ; then
+                echo " > install-docker.sh cleaned up";
+            else
+                echo " /!\ Failed to remove install-docker.sh";
+            fi
         else
-        	echo " > failed to download install-docker.sh";
-        fi
-
-        echo " > Making install-docker.sh executable";
-        if chmod +x ./install-docker.sh ; then
-        	echo " > install-docker.sh now executable";
-        else
-        	echo " /!\ install-docker.sh failed to change permission";
-        fi
-
-        echo " > Executing install-docker.sh";
-        if ./install-docker.sh ; then
-        	echo " > install-docker.sh complete";
-        else 
-        	echo " /!\ install-docker.sh failed to run correctly";
-        fi
-
-        echo " > Cleaning up install-docker.sh";
-        if rm -f ./install-docker.sh ; then
-        	echo " > install-docker.sh cleaned up";
-        else
-        	echo " /!\ Failed to remove install-docker.sh";
+            echo " /!\ Docker is already installed. Please remove before trying again."
         fi
 
     else
